+++ playground/js.play.ts
@@ -1,0 +1,8 @@
+const obj = {
+  'hello': 0,
+  'world': 1,
+}
+
+for(let item in obj) {
+  console.log(item)
+}