+++ playground/regexp.play.ts
@@ -1,0 +1,6 @@
+const png = 'image/png'
+const jpg = 'image/jpeg'
+const gif = 'image/gif'
+
+console.log('image: ', /image/.test(png))
+console.log('png|jpeg|gif: ', /png|jpeg|gif/.test(png))