+++ src/interface.ts
@@ -23,6 +23,11 @@ export interface IUserInput extends IAuthInput {
   email: string
 }
 
+export interface INote {
+  uuid: string
+  content: string
+}
+
 
 export interface IDecodedToken {
   id: string
@@ -32,4 +37,36 @@ export interface IDecodedToken {
   exp: number   // 만료시간
   iss: string   // 토큰 발급자
   sub: string   // 토큰 제목
+}
+
+
+export interface IHeadline {
+  uuid: string
+  level: number
+  hash: string
+  headline: string
+  fullHeadline: string
+  content: string
+}
+
+export interface INoteInput {
+  uuid: string
+  content?: string
+  tags?: string[]
+  origin?: string
+}
+
+export interface INoteSaving {
+  content: string
+  tags: string[]
+  origin: string
+  images: {}
+  created_at: number
+  updated_at: number
+}
+
+export interface INoteOutput extends INoteSaving {
+  uuid: string
+  heads: IHeadline[]
+  url: string
 }